+++ BulkAnimalPurchase/release-notes.md
@@ -2,6 +2,13 @@
 
 # Release notes
 
+## 0.2.2-unofficial.2-mouahrara
+Released on 6 July, 2024, for SMAPI 4.0.0 or later.
+* Added **Initial Friendship** option ✨
+* Added **Adult Animals** option ✨
+* Added **Price related** options ✨
+* Added **Incubation time related** options ✨
+
 ## 0.2.2-unofficial.1-mouahrara
 Released on 4 April, 2024, for SMAPI 4.0.0 or later.
 * Migrated to 1.6.0 and project cleanup 🚀